Refactor content ID types to long, update date format in requests/responses,...

See merge request summer20241/summer251/se1941-jv/g1/personalized-learning-system!15

+++ src/main/java/vn/fu_ohayo/controller/ContentSpeakingController.java
@@ -45,7 +45,7 @@ public class ContentSpeakingController {
     }
 
     @GetMapping("/{id}")
-    public ApiResponse<ContentSpeaking> getContentSpeaking(@PathVariable  int id) {
+    public ApiResponse<ContentSpeaking> getContentSpeaking(@PathVariable  long id) {
         return  ApiResponse.<ContentSpeaking>builder()
                 .code("200")
                 .status("success")
